refactor: outsourcing auth logic to auth hook
